backend/requirements.txt
fastapi
uvicorn[standard]
beanie
motor
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
boto3
